Renamed package name from dockerConfig to configurations
